apps/backend/requirements/base.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
django-environ==0.11.2
django-redis==5.4.0
redis==5.2.1
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
pyodbc==5.2.0
python-dateutil==2.9.0
Pillow>=10.0.0
httpx==0.27.2